Add alert before deleting medications.

+++ app/src/main/java/com/meditrack/UpdateMedicationFragment.kt
@@ -1,19 +1,18 @@
 package com.meditrack
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Button
 import android.view.*
+import android.widget.Button
 import android.widget.EditText
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.meditrack.data.DataViewModel
 import com.meditrack.data.Medication
+
 
 class UpdateMedicationFragment : Fragment() {
 
@@ -36,8 +35,14 @@ class UpdateMedicationFragment : Fragment() {
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
             R.id.action_delete -> {
-                dataViewModel.deleteMedication(args.currentItem)
-                findNavController().navigateUp()
+                AlertDialog.Builder(requireContext())
+                    .setMessage("Medication \"${args.currentItem.name}\" will be deleted permanently!")
+                    .setPositiveButton(android.R.string.ok) { _, _ ->
+                        dataViewModel.deleteMedication(args.currentItem)
+                        findNavController().navigateUp()
+                    }
+                    .setNegativeButton(android.R.string.cancel, null)
+                    .show()
             }
         }
         return super.onOptionsItemSelected(item)
